First stage of refactorings
