+++ .env.example
@@ -9,7 +9,12 @@ DEFAULT_PLANNER_MODEL="gemini-2.5-pro"
 DEFAULT_WORKER_MODEL="gemini-2.5-flash"
 DEFAULT_EVALUATOR_MODEL="gemini-2.5-pro"
 
-# LangFuse (optional, for tracing)
+# LangFuse for agent execution tracing and evaluations
 LANGFUSE_SECRET_KEY="sk-lf-..."
 LANGFUSE_PUBLIC_KEY="pk-lf-..."
 LANGFUSE_HOST="https://us.cloud.langfuse.com"
+
+# Report Generation (all optional, defaults are in implementations/report_generation/demo.py)
+REPORT_GENERATION_OUTPUT_PATH="..."
+REPORT_GENERATION_DB_PATH="..."
+REPORT_GENERATION_LANGFUSE_PROJECT_NAME="..."
