+++ implementations/report_generation/demo.py
@@ -1,94 +1,36 @@
-"""Reason-and-Act Knowledge Retrieval Agent via the OpenAI Agent SDK."""
+"""
+Demo UI for the report generation agent.
+
+Example
+-------
+$ python -m implementations.report_generation.demo
+"""
 
 import asyncio
 import logging
-import os
 from functools import partial
-from pathlib import Path
 from typing import Any, AsyncGenerator
 
 import agents
 import click
 import gradio as gr
 from aieng.agent_evals.async_client_manager import AsyncClientManager
-from aieng.agent_evals.langfuse import setup_langfuse_tracer
+from aieng.agent_evals.report_generation.agent import get_report_generation_agent
+from aieng.agent_evals.report_generation.prompts import MAIN_AGENT_INSTRUCTIONS
 from aieng.agent_evals.utils import get_or_create_session, oai_agent_stream_to_gradio_messages
 from dotenv import load_dotenv
 from gradio.components.chatbot import ChatMessage
 
-from implementations.report_generation.file_writer import get_reports_output_path, write_report_to_file
-from implementations.report_generation.prompts import MAIN_AGENT_INSTRUCTIONS
+from implementations.report_generation.env_vars import (
+    get_langfuse_project_name,
+    get_reports_output_path,
+    get_sqlite_db_path,
+)
 
 
 load_dotenv(verbose=True)
 logging.basicConfig(level=logging.INFO, format="%(asctime)s %(levelname)s %(name)s: %(message)s")
 logger = logging.getLogger(__name__)
-
-
-LANGFUSE_PROJECT_NAME = "Report Generation"
-
-
-def get_sqlite_db_path() -> Path:
-    """Get the SQLite database path.
-
-    If no path is provided in the REPORT_GENERATION_DB_PATH env var, will use the
-    default path in `implementations/report_generation/data/OnlineRetail.db`.
-
-    Returns
-    -------
-    Path
-        The SQLite database path.
-    """
-    default_sqlite_db_path = "implementations/report_generation/data/OnlineRetail.db"
-    return Path(os.getenv("REPORT_GENERATION_DB_PATH", default_sqlite_db_path))
-
-
-def get_report_generation_agent(enable_trace: bool = True) -> agents.Agent:
-    """
-    Define the report generation agent.
-
-    Parameters
-    ----------
-    enable_trace : bool, optional
-        Whether to enable tracing with Langfuse for evaluation purposes.
-        Default is True.
-
-    Returns
-    -------
-    agents.Agent
-        The report generation agent.
-    """
-    # Setup langfuse tracing if enabled
-    if enable_trace:
-        setup_langfuse_tracer(LANGFUSE_PROJECT_NAME)
-
-    # Get the client manager singleton instance
-    client_manager = AsyncClientManager.get_instance()
-
-    # Define an agent using the OpenAI Agent SDK
-    return agents.Agent(
-        name="Report Generation Agent",  # Agent name for logging and debugging purposes
-        instructions=MAIN_AGENT_INSTRUCTIONS,  # System instructions for the agent
-        # Tools available to the agent
-        # We wrap the `search_knowledgebase` method with `function_tool`, which
-        # will construct the tool definition JSON schema by extracting the necessary
-        # information from the method signature and docstring.
-        tools=[
-            agents.function_tool(
-                client_manager.sqlite_connection(get_sqlite_db_path()).execute,
-                name_override="execute_sql_query",
-                description_override="Execute a SQL query against the SQLite database.",
-            ),
-            agents.function_tool(
-                write_report_to_file,
-                description_override="Write the report data to a file.",
-            ),
-        ],
-        model=agents.OpenAIChatCompletionsModel(
-            model=client_manager.configs.default_worker_model,
-            openai_client=client_manager.openai_client,
-        ),
-    )
 
 
 async def agent_session_handler(
@@ -125,7 +67,12 @@ async def agent_session_handler(
     # previous turns in the conversation
     session = get_or_create_session(history, session_state)
 
-    main_agent = get_report_generation_agent(enable_trace=enable_trace)
+    main_agent = get_report_generation_agent(
+        instructions=MAIN_AGENT_INSTRUCTIONS,
+        sqlite_db_path=get_sqlite_db_path(),
+        reports_output_path=get_reports_output_path(),
+        langfuse_project_name=get_langfuse_project_name() if enable_trace else None,
+    )
 
     # Run the agent in streaming mode to get and display intermediate outputs
     result_stream = agents.Runner.run_streamed(main_agent, input=query, session=session)
